+++ packages/terra-docs/src/terra-dev-site/tool/example/Placeholder.jsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import { Placeholder } from '@cerner/terra-docs';
+
+export default () => <Placeholder title="Placeholder" />;