+++ packages/terra-docs/tests/wdio/placeholder-spec.js
@@ -1,0 +1,8 @@
+Terra.describeViewports('Placeholder', ['medium'], () => {
+  describe('Default Placeholder', () => {
+    it('checks accessibility', () => {
+      browser.url('/raw/tests/terra-docs/docs/placeholder');
+      Terra.validates.element();
+    });
+  });
+});