+++ packages/terra-markdown/src/terra-dev-site/test/markdown/blockquote.test.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import Markdown from '../../../../lib/Markdown';
-import markdown from './blockquote.md';
-
-const MarkdownTest = () => (
-  <Markdown src={markdown} />
-);
-
-export default MarkdownTest;