+++ packages/terra-form-textarea/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+* Changed
+  * Removed reference to terra-markdown
+
+## 5.5.0 - (September 29, 2020)
+
+* Changed
+  * Minor dependency version bump
+
 ## 5.4.0 - (September 16, 2020)
 
 * Added